[#86] feat: 로직분리
